refactor, station-management 페이지 기능 파일별로 리펙토링
